마스터에 잘못 머지한 pr revert

+++ src/main/kotlin/com/dotori/v2/global/util/ProfileImageService.kt
@@ -12,15 +12,13 @@ class ProfileImageService (
     private val s3Service: S3Service,
     private val redisCacheService: RedisCacheService
 ) {
+    fun imageUpload(member: Member, multipartFiles: MultipartFile?, isUpdate: Boolean = false) {
 
-    fun imageUpload(member: Member, multipartFiles: MultipartFile?) {
         validateExtension(multipartFiles)
 
-        val uploadFileUrl = s3Service.uploadSingleFile(multipartFiles)
+        val uploadFileUrl: String? = s3Service.uploadSingleFile(multipartFiles)
 
-        if (member.isProfileImageExists()) {
-            s3Service.deleteFile(member.profileImage!!)
-        }
+        if (isUpdate) s3Service.deleteFile(member.profileImage!!)
 
         member.updateProfileImage(uploadFileUrl)
